requirements.txt
scikit-image~=0.22.0
numpy~=1.26.1
setuptools~=68.2.0
matplotlib~=3.8.1
scipy~=1.11.3
IPython
pandas~=2.1.3
seaborn~=0.13.0
scikit-learn~=1.3.2
Pillow~=10.0.1